refactor: pages 폴더를 components 밖으로 분리
