+++ src/App.js
@@ -3,11 +3,11 @@ import { HashRouter as Router, Redirect, Route, Switch } from 'react-router-dom'
 
 import Header from './components/Header';
 import Navigation from './components/Navigation';
-import CartPage from './components/pages/CartPage';
-import CheckoutPage from './components/pages/CheckoutPage';
-import OrdersPage from './components/pages/OrdersPage';
-import ProductDetailPage from './components/pages/ProductDetailPage';
-import ProductListPage from './components/pages/ProductListPage';
+import CartPage from './pages/CartPage';
+import CheckoutPage from './pages/CheckoutPage';
+import OrdersPage from './pages/OrdersPage';
+import ProductDetailPage from './pages/ProductDetailPage';
+import ProductListPage from './pages/ProductListPage';
 import ShoppingCart from './components/common/Icon/ShoppingCart';
 
 import { HEADER, PAGES } from './constants/appInfo';
