+++ src/pages/OrdersPage/style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import PALETTE from '../../../constants/palette';
+import PALETTE from '../../constants/palette';
 
 export const OrdersPageContainer = styled.div`
   height: 100%;
