+++ src/pages/ProductDetailPage/style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import PALETTE from '../../../constants/palette';
+import PALETTE from '../../constants/palette';
 
 export const Container = styled.div`
   width: 45%;
